# DOCUMENTATION/openapi.yaml
openapi: "3.0.0"
info:
  title: RENTVERSE API v2.4
  description: A comprehensive API for the RENTVERSE platform, featuring a secure refresh token authentication mechanism and an expanded feature set including geolocation, property views, and automated discovery of nearby facilities.
  version: 2.4.0
servers:
  - url: http://localhost:8080/api
    description: Local Development Server via API Gateway
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication]
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Log in a user and receive tokens
      tags: [Authentication]
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLogin'
      responses:
        '200':
          description: Login successful. Returns an access token in the body and a refresh token in a secure, HttpOnly cookie.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: refreshToken=abcde12345; HttpOnly; Path=/; Max-Age=604800
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Obtain a new access token using a refresh token
      tags: [Authentication]
      operationId: refreshToken
      description: This endpoint relies on the HttpOnly refresh token cookie sent by the browser. No request body is needed.
      responses:
        '200':
          description: A new access token is issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid or expired refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Log out a user
      tags: [Authentication]
      operationId: logoutUser
      description: Invalidates the user's refresh token by deleting it from the database and clearing the cookie.
      responses:
        '204':
          description: Logout successful. No content returned.
        '401':
          description: No active session to log out from.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserRegister:
      type: object
      required: [fullName, email, password, role]
      properties:
        fullName: { type: string, example: "Budi Siregar" }
        email: { type: string, format: email, example: "budi.siregar@example.com" }
        password: { type: string, format: password, example: "SecurePassword123" }
        role: { type: string, enum: [PROPERTY_OWNER, TENANT] }
    UserLogin:
      type: object
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
    UserResponse:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        phone: { type: string, nullable: true }
        country: { type: string, nullable: true }
        profilePictureUrl: { type: string, nullable: true }
        role: { type: string, enum: [PROPERTY_OWNER, TENANT, ADMIN] }
...
    ErrorResponse:
      type: object
      properties:
        message: { type: string }